+++ src/ggg/feature_variations.zig
@@ -2,7 +2,7 @@ const lib = @import("../lib.zig");
 const parser = @import("../parser.zig");
 const utils = @import("../utils.zig");
 
-pub const VariationIndex = lib.opentype_layout.VariationIndex;
+const otl = lib.opentype_layout;
 const LookupList = @import("lookup.zig").LookupList;
 
 /// A [Feature Variations Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#featurevariations-table).
@@ -31,8 +31,8 @@ pub fn parse(
 pub fn find_index(
     self: *const FeatureVariations,
     coords: []const lib.NormalizedCoordinate,
-) ?VariationIndex {
-    var i: VariationIndex = 0;
+) ?otl.VariationIndex {
+    var i: otl.VariationIndex = 0;
     while (i < self.records.len()) : (i += 1) {
         const record = self.records.get(i) orelse return null;
         const offset: usize = record.conditions[0];
@@ -45,6 +45,21 @@ pub fn find_index(
     }
 
     return null;
+}
+
+/// Returns a [`Feature`] at specified indices.
+pub fn find_substitute(
+    self: *const FeatureVariations,
+    feature_index: otl.FeatureIndex,
+    variation_index: otl.VariationIndex,
+) ?otl.Feature {
+    const record = self.records.get(variation_index) orelse return null;
+    const offset = record.substitutions[0];
+
+    const data = utils.slice(self.data, offset) catch return null;
+    const subst = FeatureTableSubstitution.parse(data) catch return null;
+
+    return subst.find_substitute(feature_index);
 }
 
 const ConditionSet = struct {
@@ -123,6 +138,64 @@ const FeatureVariationRecord = struct {
     pub const FromData = struct {
         // [ARS] impl of FromData trait
         pub const SIZE: usize = 8;
+
+        pub fn parse(
+            data: *const [SIZE]u8,
+        ) parser.Error!Self {
+            return try parser.parse_struct_from_data(Self, data);
+        }
+    };
+};
+
+const FeatureTableSubstitution = struct {
+    data: []const u8,
+    records: parser.LazyArray16(FeatureTableSubstitutionRecord),
+
+    fn parse(
+        data: []const u8,
+    ) parser.Error!FeatureTableSubstitution {
+        var s = parser.Stream.new(data);
+        const major_version = try s.read(u16);
+        s.skip(u16); // minor version
+        if (major_version != 1) return error.DataError;
+
+        const count = try s.read(u16);
+        const records = try s.read_array(FeatureTableSubstitutionRecord, count);
+        return .{
+            .data = data,
+            .records = records,
+        };
+    }
+
+    fn find_substitute(
+        self: FeatureTableSubstitution,
+        feature_index: otl.FeatureIndex,
+    ) ?otl.Feature {
+        var iter = self.records.iterator();
+        while (iter.next()) |record| {
+            if (record.feature_index == feature_index) {
+                const offset = record.feature[0];
+
+                const data = utils.slice(self.data, offset) catch return null;
+                // TODO: set tag
+                return otl.Feature.parse(
+                    .from_bytes("DFLT"),
+                    data,
+                ) catch null;
+            }
+        }
+        return null;
+    }
+};
+
+const FeatureTableSubstitutionRecord = struct {
+    feature_index: otl.FeatureIndex,
+    feature: parser.Offset32,
+
+    const Self = @This();
+    pub const FromData = struct {
+        // [ARS] impl of FromData trait
+        pub const SIZE: usize = 6;
 
         pub fn parse(
             data: *const [SIZE]u8,